Move Base UI page generation to project-specific directory
