+++ packages/api-docs-builder/projects/baseUi/projectSettings.ts
@@ -3,6 +3,7 @@ import { ProjectSettings } from 'buildApi';
 import findApiPages from '../../utils/findApiPages';
 import getBaseUiComponentInfo from './getBaseUiComponentInfo';
 import getBaseUiHookInfo from './getBaseUiHookInfo';
+import generateBaseUIApiPages from './generateBaseUiApiPages';
 
 const projectSettings: ProjectSettings = {
   output: {
@@ -12,6 +13,9 @@ const projectSettings: ProjectSettings = {
   getApiPages: () => findApiPages('docs/pages/base-ui/api'),
   getComponentInfo: getBaseUiComponentInfo,
   getHookInfo: getBaseUiHookInfo,
+  onCompleted: () => {
+    generateBaseUIApiPages();
+  },
 };
 
 export default projectSettings;
